requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
azure-common==1.1.28
azure-core==1.38.2
azure-search-documents==11.6.0
certifi==2026.2.25
charset-normalizer==3.4.5
distro==1.9.0
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
Markdown==3.10.2
openai==2.26.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.8.0
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
